src/requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiogram_dialog==2.3.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==25.1.0
bcrypt==4.0.1
cachetools==5.5.1
certifi==2024.12.14
click==8.1.8
fastapi==0.115.8
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
hiredis==3.1.0
idna==3.10
Jinja2==3.1.5
magic-filter==1.0.12
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
passlib==1.7.4
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
redis==5.2.1
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
yarl==1.18.3
